const_oid

Module arcs

Arcs are integer values which exist within an OID’s hierarchy.

Structs

Arcs

Iterator over Arc values (a.k.a. nodes) in an ObjectIdentifier.

RootArcs 🔒

Byte containing the first and second arcs of an OID.

Constants

ARC_MAX_BYTES 🔒

Maximum number of bytes supported in an arc.

ARC_MAX_FIRST 🔒

Maximum value of the first arc in an OID.

ARC_MAX_LAST_OCTET 🔒

Maximum value of the last byte in an arc.

ARC_MAX_SECOND 🔒

Maximum value of the second arc in an OID.

Type Aliases

Arc

Type alias used to represent an “arc”, i.e. integer identifier value, where an OID comprises a sequence of arcs.