const_oid::db

Struct Database

pub struct Database<'a>(&'a [(&'a ObjectIdentifier, &'a str)]);

A query interface for OIDs/Names.

Tuple Fields
0: &'a [(&'a ObjectIdentifier, &'a str)]
Implementations

impl<'a> Database<'a>

pub fn resolve<'b>(&self, oid: &'b str) -> Result<&'b str, Error>
where 'a: 'b,

Looks up a name for an OID.

Errors if the input is not a valid OID. Returns the input if no name is found.

pub const fn by_oid(&self, oid: &ObjectIdentifier) -> Option<&'a str>

Finds a named oid by its associated OID.

pub const fn by_name(&self, name: &str) -> Option<&'a ObjectIdentifier>

Finds a named oid by its associated name.

pub const fn find_names_for_oid(&self, oid: ObjectIdentifier) -> Names<'a>

Return the list of matched name for the OID.