Struct const_oid::ObjectIdentifier

pub struct ObjectIdentifier { /* private fields */ }

Object identifier (OID).

OIDs are hierarchical structures consisting of “arcs”, i.e. integer identifiers.

Validity

In order for an OID to be considered valid by this library, it must meet the following criteria:

The OID MUST have at least 3 arcs

The first arc MUST be within the range 0-2

The second arc MUST be within the range 0-39

The BER/DER encoding of the OID MUST be shorter than ObjectIdentifier::MAX_SIZE

Implementations

impl ObjectIdentifier

pub const MAX_SIZE: usize = 39usize

Maximum size of a BER/DER-encoded OID in bytes.

pub const fn new(s: &str) -> Self

Parse an ObjectIdentifier from the dot-delimited string form, e.g.:

use const_oid::ObjectIdentifier;

pub const MY_OID: ObjectIdentifier = ObjectIdentifier::new("1.2.840.113549.1.1.1");

Panics

This method panics in the event the OID is malformed according to the “Validity” rules given in the toplevel documentation for this type.

For that reason this method is ONLY recommended for use in constants (where it will generate a compiler error instead).

To parse an OID from a &str slice fallibly and without panicking, use the FromStr impl instead (or via str’s parse method).

pub fn from_arcs(arcs: &[Arc]) -> Result<Self>

Parse an OID from a slice of Arc values (i.e. integers).

pub fn from_bytes(ber_bytes: &[u8]) -> Result<Self>

Parse an OID from from its BER/DER encoding.

pub fn as_bytes(&self) -> &[u8]

Get the BER/DER serialization of this OID as bytes.

Note that this encoding omits the tag/length, and only contains the value portion of the encoded OID.

pub fn arc(&self, index: usize) -> Option<Arc>

Return the arc with the given index, if it exists.

pub fn arcs(&self) -> Arcs<'_>

Iterate over the arcs (a.k.a. nodes) of an ObjectIdentifier.

Returns Arcs, an iterator over Arc values representing the value of each arc/node.