const_oid

Struct ObjectIdentifierRef

pub struct ObjectIdentifierRef { /* private fields */ }

OID reference type: wrapper for the BER serialization.

Implementations

impl ObjectIdentifierRef

pub fn from_bytes(ber: &[u8]) -> Result<&Self>

Create an ObjectIdentifierRef, validating that the provided byte slice contains a valid BER/DER encoding.

pub const fn as_bytes(&self) -> &[u8]

Get the BER/DER serialization of this OID as bytes.

Note that this encoding omits the ASN.1 tag/length, and only contains the value portion of the encoded OID.

pub fn arc(&self, index: usize) -> Option<Arc>

Return the arc with the given index, if it exists.

pub fn arcs(&self) -> Arcs<'_>

Iterate over the arcs (a.k.a. nodes) of an ObjectIdentifier.

Returns Arcs, an iterator over Arc values.

pub fn len(&self) -> usize

Get the length of this ObjectIdentifier in arcs.