Crate const_oid

Const-friendly ISO/IEC Object Identifier (OID) library with support for heapless no_std-environments.

About OIDs

Object Identifiers, a.k.a. OIDs, are an International Telecommunications Union (ITU) and ISO/IEC standard for naming any object, concept, or "thing" with a globally unambiguous persistent name.

The following is an example of an OID, in this case identifying the rsaEncryption algorithm:

1.2.840.113549.1.1.1

For more information, see: https://en.wikipedia.org/wiki/Object_identifier

Minimum Supported Rust Version

This crate requires Rust 1.46 at a minimum.

Minimum supported Rust version may be changed in the future, but it will be accompanied with a minor version bump.

Structs

Error
Error type

ObjectIdentifier
Object identifier (OID)

Constants

MAX_NODES
Maximum number of nodes in an OID.

Type Definitions

Result
Result type