const_oid

Module db

OID Names Database

The contents of this database are generated from the official IANA Object Identifier Descriptors Registry CSV file and from RFC 5280. If we are missing values you care about, please contribute a patch to oiddbgen (a subcrate in the source code) to generate the values from the relevant standard.

Modules

gen 🔒

!! DO NOT EDIT !!: This file is auto-generated by oiddbgen.

rfc1274

rfc2079

rfc2164

rfc2247

rfc2252

rfc2256

rfc2293

rfc2589

rfc2739

rfc2798

rfc3280

rfc3296

rfc3671

rfc3672

rfc3687

rfc3698

rfc3703

rfc3712

rfc4104

rfc4237

rfc4403

rfc4512

rfc4517

rfc4519

rfc4523

rfc4524

rfc4530

rfc4876

rfc5020

rfc5280

rfc5911

rfc5912

rfc6109

rfc6268

rfc6960

rfc6962

rfc7107

rfc7299

rfc7532

rfc7612

rfc8284

rfc8410

rfc8894

Structs

Database

A query interface for OIDs/Names.

Names

Iterator returning the multiple names that may be associated with an OID.

Constants

DB

Functions

eq 🔒

A const implementation of byte equals.

eq_case 🔒

A const implementation of case-insensitive ASCII equals.