Struct const_oid::ObjectIdentifier

pub struct ObjectIdentifier { /* private fields */ }

Object identifier (OID).

OIDs are hierarchical structures consisting of “arcs”, i.e. integer identifiers.

Validity

In order for an OID to be considered valid by this library, it must meet the following criteria:

The OID MUST have at least 3 arcs

The first arc MUST be within the range 0-2

The second arc MUST be within the range 0-39

The BER/DER encoding of the OID MUST be shorter than ObjectIdentifier::MAX_SIZE

Implementations

impl ObjectIdentifier

pub const MAX_SIZE: usize = 39usize

Maximum size of a BER/DER-encoded OID in bytes.

pub const fn new_unwrap(s: &str) -> Self

Parse an ObjectIdentifier from the dot-delimited string form, panicking on parse errors.

This function exists as a workaround for unwrap not yet being stable in const fn contexts, and is intended to allow the result to be bound to a constant value:

use const_oid::ObjectIdentifier;

pub const MY_OID: ObjectIdentifier = ObjectIdentifier::new_unwrap("1.2.840.113549.1.1.1");

In future versions of Rust it should be possible to replace this with ObjectIdentifier::new(...).unwrap().

Use ObjectIdentifier::new for fallible parsing.

pub const fn new(s: &str) -> Result<Self>

Parse an ObjectIdentifier from the dot-delimited string form.

pub fn from_arcs(arcs: impl IntoIterator<Item = Arc>) -> Result<Self>

Parse an OID from a slice of Arc values (i.e. integers).

pub fn from_bytes(ber_bytes: &[u8]) -> Result<Self>

Parse an OID from from its BER/DER encoding.

pub fn as_bytes(&self) -> &[u8]

Get the BER/DER serialization of this OID as bytes.

Note that this encoding omits the tag/length, and only contains the value portion of the encoded OID.

pub fn arc(&self, index: usize) -> Option<Arc>

Return the arc with the given index, if it exists.

pub fn arcs(&self) -> Arcs<'_>

Iterate over the arcs (a.k.a. nodes) of an ObjectIdentifier.

Returns Arcs, an iterator over Arc values.

pub fn len(&self) -> usize

Get the length of this ObjectIdentifier in arcs.

pub fn parent(&self) -> Option<Self>

Get the parent OID of this one (if applicable).

pub const fn push_arc(self, arc: Arc) -> Result<Self>

Push an additional arc onto this OID, returning the child OID.